+++ backend/src/service/clip/picker.rs
@@ -1,33 +1,24 @@
 use std::collections::hash_map::Entry;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 
-use clip_mash_types::{
-    Clip, EqualLengthClipOptions, MarkerId, PmvClipOptions, RoundRobinClipOptions,
-    WeightedRandomClipOptions,
-};
+use clip_mash_types::MarkerId;
 use float_cmp::approx_eq;
-use rand::distributions::WeightedIndex;
-use rand::prelude::Distribution;
 use rand::rngs::StdRng;
-use rand::seq::{IteratorRandom, SliceRandom};
-use tracing::{debug, info};
+use rand::seq::IteratorRandom;
 
-use super::pmv::PmvClipLengths;
 use crate::service::Marker;
 
-const MIN_DURATION: f64 = 1.5;
-
 #[derive(Debug)]
-struct MarkerStart {
-    start_time: f64,
-    end_time: f64,
-    index: usize,
+pub struct MarkerStart {
+    pub start_time: f64,
+    pub end_time: f64,
+    pub index: usize,
 }
 
 #[derive(Debug)]
-struct MarkerState {
-    data: HashMap<i64, MarkerStart>,
-    markers: Vec<Marker>,
+pub struct MarkerState {
+    pub data: HashMap<i64, MarkerStart>,
+    pub markers: Vec<Marker>,
 }
 
 impl MarkerState {
@@ -69,7 +60,7 @@ impl MarkerState {
         }
     }
 
-    fn find_marker_by_index(&self, index: usize) -> Option<Marker> {
+    pub fn find_marker_by_index(&self, index: usize) -> Option<Marker> {
         if self.markers.is_empty() {
             None
         } else {
@@ -77,265 +68,12 @@ impl MarkerState {
         }
     }
 
-    fn find_marker_by_title(&self, title: &str, rng: &mut StdRng) -> Option<Marker> {
+    pub fn find_marker_by_title(&self, title: &str, rng: &mut StdRng) -> Option<Marker> {
         self.markers
             .iter()
             .filter(|m| &m.title == title)
             .choose(rng)
             .cloned()
-    }
-}
-
-pub trait ClipPicker {
-    type Options;
-
-    fn pick_clips(
-        &mut self,
-        markers: Vec<Marker>,
-        options: Self::Options,
-        rng: &mut StdRng,
-    ) -> Vec<Clip>;
-}
-
-pub struct RoundRobinClipPicker;
-
-impl ClipPicker for RoundRobinClipPicker {
-    type Options = RoundRobinClipOptions;
-
-    fn pick_clips(
-        &mut self,
-        markers: Vec<Marker>,
-        options: Self::Options,
-        rng: &mut StdRng,
-    ) -> Vec<Clip> {
-        info!("using RoundRobinClipPicker to make clips from markers {markers:#?} with options {options:#?}");
-
-        let max_duration = options.length;
-        let mut total_duration = 0.0;
-        let mut clips = vec![];
-        let mut marker_idx = 0;
-        let has_music = matches!(options.clip_lengths, PmvClipOptions::Songs(_));
-        let mut clip_lengths: PmvClipLengths = options.clip_lengths.into();
-        let mut marker_state = MarkerState::new(markers);
-
-        while total_duration <= options.length {
-            debug!("marker state: {marker_state:#?}, total duration: {total_duration}, target duration: {}", options.length);
-            if marker_state.markers.is_empty() {
-                info!("no more markers to pick from, stopping");
-                break;
-            }
-
-            if let Some(marker) = marker_state.find_marker_by_index(marker_idx) {
-                if let Some(MarkerStart {
-                    start_time: start,
-                    index,
-                    ..
-                }) = marker_state.get(&marker.id)
-                {
-                    let clip_duration = clip_lengths.pick_duration(rng);
-                    if clip_duration.is_none() {
-                        break;
-                    }
-                    let clip_duration = clip_duration.unwrap();
-                    let end = (start + clip_duration).min(marker.end_time);
-                    let duration = end - start;
-                    if has_music || duration >= MIN_DURATION {
-                        info!(
-                            "adding clip for video {} with duration {duration} and title {}",
-                            marker.video_id, marker.title
-                        );
-                        clips.push(Clip {
-                            index_within_marker: *index,
-                            index_within_video: marker.index_within_video,
-                            marker_id: marker.id,
-                            range: (*start, end),
-                            source: marker.video_id.source(),
-                            video_id: marker.video_id.clone(),
-                        });
-                    }
-
-                    total_duration += duration;
-                    marker_idx += 1;
-                    marker_state.update(&marker.id, end, index + 1);
-                }
-            }
-        }
-
-        let clips_duration: f64 = clips.iter().map(|c| c.duration()).sum();
-        if clips_duration > max_duration {
-            let slack = (clips_duration - max_duration) / clips.len() as f64;
-            info!("clip duration {clips_duration} longer than permitted maximum duration {max_duration}, making each clip {slack} shorter");
-            for clip in &mut clips {
-                clip.range.1 = clip.range.1 - slack;
-            }
-        }
-
-        clips
-    }
-}
-
-pub struct WeightedRandomClipPicker;
-
-fn validate_options(
-    markers: &[Marker],
-    options: &WeightedRandomClipOptions,
-    weight_labels: &HashSet<&str>,
-) {
-    for (title, weight) in &options.weights {
-        assert!(
-            *weight > 0.0,
-            "weight for title {} must be greater than 0",
-            title
-        );
-        let marker_count = markers.iter().filter(|m| &m.title == title).count();
-        assert!(marker_count > 0, "no markers found for title {}", title);
-    }
-
-    let weights_exist = markers
-        .iter()
-        .all(|m| weight_labels.contains(m.title.as_str()));
-    assert!(weights_exist, "all markers must have a weight");
-}
-
-impl ClipPicker for WeightedRandomClipPicker {
-    type Options = WeightedRandomClipOptions;
-
-    fn pick_clips(
-        &mut self,
-        mut markers: Vec<Marker>,
-        mut options: Self::Options,
-        rng: &mut StdRng,
-    ) -> Vec<Clip> {
-        info!("using WeightedRandomClipPicker to make clips: {options:#?}");
-        debug!("using markers: {markers:#?}");
-        options.weights.retain(|(_, weight)| *weight > 0.0);
-        let weight_labels: HashSet<_> = options
-            .weights
-            .iter()
-            .map(|(label, _)| label.as_str())
-            .collect();
-        markers.retain(|m| weight_labels.contains(m.title.as_str()));
-
-        validate_options(&markers, &options, &weight_labels);
-        let choices = options.weights;
-
-        let distribution = WeightedIndex::new(choices.iter().map(|item| item.1))
-            .expect("could not build distribution");
-        let mut total_duration = 0.0;
-        let mut marker_state = MarkerState::new(markers);
-        let mut clips = vec![];
-        let mut clip_lengths: PmvClipLengths = options.clip_lengths.into();
-
-        while total_duration <= options.length {
-            debug!("marker state: {marker_state:#?}, total duration: {total_duration}, target duration: {}", options.length);
-            if marker_state.markers.is_empty() {
-                info!("no more markers to pick from, stopping");
-                break;
-            }
-            let marker_tag = &choices[distribution.sample(rng)].0;
-            if let Some(marker) = marker_state.find_marker_by_title(&marker_tag, rng) {
-                let clip_duration = clip_lengths.pick_duration(rng);
-                if clip_duration.is_none() {
-                    break;
-                }
-                if let Some(MarkerStart {
-                    start_time: start,
-                    index,
-                    ..
-                }) = marker_state.get(&marker.id)
-                {
-                    let clip_duration = clip_duration.unwrap();
-                    let end = (start + clip_duration).min(marker.end_time);
-                    let duration = end - start;
-
-                    clips.push(Clip {
-                        index_within_marker: *index,
-                        index_within_video: marker.index_within_video,
-                        marker_id: marker.id,
-                        range: (*start, end),
-                        source: marker.video_id.source(),
-                        video_id: marker.video_id.clone(),
-                    });
-                    info!(
-                        "adding clip for video {} with duration {duration} and title {}",
-                        marker.video_id, marker.title
-                    );
-
-                    marker_state.update(&marker.id, end, index + 1);
-                    total_duration += duration;
-                }
-            } else {
-                debug!(
-                    "no marker found for title {marker_tag}, skipping, remaining markers: {:?}",
-                    marker_state.markers
-                );
-            }
-        }
-        let clips_duration: f64 = clips.iter().map(|c| c.duration()).sum();
-        if clips_duration > options.length {
-            let slack = (clips_duration - options.length) / clips.len() as f64;
-            info!("clip duration {clips_duration} longer than permitted maximum duration {}, making each clip {slack} shorter", options.length);
-            for clip in &mut clips {
-                clip.range.1 = clip.range.1 - slack;
-            }
-        }
-
-        clips
-    }
-}
-
-pub struct EqualLengthClipPicker;
-
-impl ClipPicker for EqualLengthClipPicker {
-    type Options = EqualLengthClipOptions;
-
-    fn pick_clips(
-        &mut self,
-        markers: Vec<Marker>,
-        options: Self::Options,
-        rng: &mut StdRng,
-    ) -> Vec<Clip> {
-        assert!(options.divisors.len() > 0, "divisors must not be empty");
-        info!("using EqualLengthClipPicker to make clips: {options:?}");
-
-        let duration = options.clip_duration;
-        let clip_lengths: Vec<f64> = options
-            .divisors
-            .into_iter()
-            .map(|d| (duration / d).max(MIN_DURATION))
-            .collect();
-        let mut clips = vec![];
-        for marker in markers {
-            let start = marker.start_time;
-            let end = marker.end_time;
-
-            let mut index = 0;
-            let mut offset = start;
-            while offset < end {
-                let duration = clip_lengths.choose(rng).unwrap();
-                let start = offset;
-                let end = (offset + duration).min(end);
-                let duration = end - start;
-                if duration > MIN_DURATION {
-                    info!(
-                        "adding clip for video {} with duration {duration} and title {}",
-                        marker.video_id, marker.title
-                    );
-                    clips.push(Clip {
-                        source: marker.video_id.source(),
-                        video_id: marker.video_id.clone(),
-                        marker_id: marker.id,
-                        range: (start, end),
-                        index_within_marker: index,
-                        index_within_video: marker.index_within_video,
-                    });
-                    index += 1;
-                }
-                offset += duration;
-            }
-        }
-
-        clips
     }
 }
 
@@ -349,9 +87,10 @@ mod tests {
     use float_cmp::assert_approx_eq;
     use tracing_test::traced_test;
 
-    use super::{validate_options, RoundRobinClipPicker};
-    use crate::service::clip::picker::{ClipPicker, WeightedRandomClipPicker};
-    use crate::service::fixtures::{self, other_markers};
+    use crate::service::clip::round_robin::RoundRobinClipPicker;
+    use crate::service::clip::weighted::WeightedRandomClipPicker;
+    use crate::service::clip::ClipPicker;
+    use crate::service::fixtures;
     use crate::util::create_seeded_rng;
 
     #[traced_test]
@@ -428,107 +167,5 @@ mod tests {
         let clip_duration: f64 = clips.iter().map(|c| c.duration()).sum();
 
         assert_approx_eq!(f64, clip_duration, video_duration, epsilon = 0.1);
-    }
-
-    #[traced_test]
-    #[test]
-    fn test_weighted_marker_infinite_loop_bug() {
-        let options = WeightedRandomClipOptions {
-            weights: vec![
-                ("Cowgirl".into(), 1.0),
-                ("Doggy Style".into(), 1.0),
-                ("Handjiob".into(), 1.0),
-                ("Mating Press".into(), 1.0),
-                ("Missionary".into(), 0.0),
-                ("Sex".into(), 1.0),
-                ("Sideways".into(), 1.0),
-            ],
-            length: 956.839832,
-            clip_lengths: PmvClipOptions::Randomized(RandomizedClipOptions {
-                base_duration: 30.0,
-                divisors: vec![2.0, 3.0, 4.0],
-            }),
-        };
-        let markers = other_markers();
-        let mut rng = create_seeded_rng(None);
-        let mut picker = WeightedRandomClipPicker;
-        let clips = picker.pick_clips(markers, options, &mut rng);
-        let clip_duration: f64 = clips.iter().map(|c| c.duration()).sum();
-        assert!(clip_duration >= 0.0);
-    }
-
-    #[test]
-    fn test_validate_options_valid() {
-        let markers = vec![
-            fixtures::create_marker("A", 0.0, 30.0, 0),
-            fixtures::create_marker("B", 0.0, 30.0, 1),
-            fixtures::create_marker("C", 0.0, 30.0, 2),
-        ];
-        let options = WeightedRandomClipOptions {
-            weights: vec![
-                ("A".to_string(), 1.0),
-                ("B".to_string(), 2.0),
-                ("C".to_string(), 3.0),
-            ],
-            clip_lengths: PmvClipOptions::Randomized(RandomizedClipOptions {
-                base_duration: 30.0,
-                divisors: vec![2.0, 3.0, 4.0],
-            }),
-            length: 30.0,
-        };
-        let weight_labels = vec!["A", "B", "C"].into_iter().collect();
-
-        validate_options(&markers, &options, &weight_labels);
-    }
-
-    #[test]
-    #[should_panic(expected = "weight for title A must be greater than 0")]
-    fn test_validate_options_zero_weight() {
-        let markers = vec![fixtures::create_marker("A", 0.0, 30.0, 0)];
-        let options = WeightedRandomClipOptions {
-            weights: vec![("A".to_string(), 0.0)],
-            clip_lengths: PmvClipOptions::Randomized(RandomizedClipOptions {
-                base_duration: 30.0,
-                divisors: vec![2.0, 3.0, 4.0],
-            }),
-            length: 30.0,
-        };
-        let weight_labels = vec!["A"].into_iter().collect();
-
-        validate_options(&markers, &options, &weight_labels);
-    }
-
-    #[test]
-    #[should_panic(expected = "no markers found for title B")]
-    fn test_validate_options_missing_marker() {
-        let markers = vec![fixtures::create_marker("A", 0.0, 30.0, 0)];
-        let options = WeightedRandomClipOptions {
-            weights: vec![("B".to_string(), 1.0)],
-            clip_lengths: PmvClipOptions::Randomized(RandomizedClipOptions {
-                base_duration: 30.0,
-                divisors: vec![2.0, 3.0, 4.0],
-            }),
-            length: 30.0,
-        };
-        let weight_labels = vec!["B"].into_iter().collect();
-
-        validate_options(&markers, &options, &weight_labels);
-    }
-
-    #[test]
-    #[should_panic(expected = "all markers must have a weight")]
-    fn test_validate_options_missing_weight() {
-        let markers = vec![fixtures::create_marker("A", 0.0, 30.0, 0)];
-        let options = WeightedRandomClipOptions {
-            weights: vec![],
-            clip_lengths: PmvClipOptions::Randomized(RandomizedClipOptions {
-                base_duration: 30.0,
-                divisors: vec![2.0, 3.0, 4.0],
-            }),
-            length: 30.0,
-        };
-        let weight_labels = vec![].into_iter().collect();
-
-        validate_options(&markers, &options, &weight_labels);
     }
 }
